Time: 245 ms (72.18%), Space: 99.9 MB (31.79%) - LeetHub

+++ 2300-successful-pairs-of-spells-and-potions/2300-successful-pairs-of-spells-and-potions.cpp
@@ -20,7 +20,7 @@ public:
         while(low<high)
         {
             int mid=(low+high)/2;
-            if(static_cast<long long>(potions[mid]*(spell)>=success))
+            if((potions[mid]*(spell)>=success))
             {
                 high=mid;
             }
